requirements.txt
tensorflow>=2.10.0
streamlit>=1.25.0
numpy>=1.23.0
Pillow>=9.4.0
python-dotenv>=1.0.0  # (Optional for environment variables)
tensorflow>=2.10.0
gradio>=3.50.0,<4.0.0  # Pinning to stable 3.x version
numpy>=1.23.0
Pillow>=9.4.0
tensorflow>=2.10.0
gradio==3.50.2  # Pinned version to prevent new issues
numpy>=1.23.0
Pillow>=9.4.0